fix change theme
